+++ examples/bookstore/src/document.rs
@@ -2,23 +2,19 @@ use driver::document::DocumentModel;
 
 /// Displays a document listing its data.
 pub fn display_document(document: &DocumentModel) {
-    println!("_id: {}", document.id());
+    println!("{{");
+    println!("  _id: {}", document.id());
 
     for (key, value) in &document.data {
-        println!("{}: {}", key, value);
+        println!("  {}: {}", key, value);
     }
+
+    println!("}}");
 }
 
 /// Displays a list of documents and their data.
 pub fn display_document_list(documents: &Vec<DocumentModel>) {
     for document in documents {
-        println!("{{");
-        println!("  _id: {}", document.id);
-
-        for (key, value) in &document.data {
-            println!("  {}: {}", key, value);
-        }
-
-        println!("}}");
+        display_document(document);
     }
 }
